Fix: split non-component exports out of ui files for fast refresh
All three react-refresh/only-export-components warnings came from
shadcn files exporting a component plus a shared value, which silently
downgraded HMR to full reloads for Button/Toggle/ThemeProvider edits:

- button.tsx exported buttonVariants that nothing imports - dropped.
- toggleVariants moves to toggle-variants.ts (toggle-group.tsx was its
  only consumer). That orphaned toggle.tsx entirely - the Toggle
  component itself was never used, only its cva def - so the file and
  the @radix-ui/react-toggle dependency go too (knip caught both as
  new findings once the variants moved out).
- useTheme + ThemeProviderContext move to theme-context.ts, mirroring
  the DomainSettingsProvider / domain-settings-context split.

eslint is now fully clean; knip drops below its previous baseline.
