+++ src/components/ui/theme-context.ts
@@ -1,0 +1,30 @@
+import { createContext, useContext } from "react";
+
+export type Theme = "dark" | "light" | "system";
+
+type ThemeProviderState = {
+  theme: Theme;
+  setTheme: (theme: Theme) => void;
+};
+
+const initialState: ThemeProviderState = {
+  theme: "system",
+  setTheme: () => null,
+};
+
+/**
+ * Context + consumer hook, split from `theme-provider.tsx` so that file
+ * only exports the component (react-refresh fast-refresh requirement) —
+ * same pattern as `contexts/domain-settings-context.ts`.
+ */
+export const ThemeProviderContext =
+  createContext<ThemeProviderState>(initialState);
+
+export const useTheme = () => {
+  const context = useContext(ThemeProviderContext);
+
+  if (context === undefined)
+    throw new Error("useTheme must be used within a ThemeProvider");
+
+  return context;
+};
